+++ CHANGELOG.rst
@@ -8,6 +8,8 @@ Changelog
 * Add endless pagination on requirements file, used on opps editor for locate image
 * Add missing Notification model migration
 * Update README file, add noimput in syncdb run
+* Add new field recursive on Boxe (app) QuerySet, #320
+* Fixed version dependency Pillow/Django-Filebrowser, #322
 
 
 0.2.3